e2e/requirements.txt
aiohttp==3.9.3
aiohttp-cors==0.7.0
aiorwlock==1.4.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
blessed==1.20.0
boto3==1.34.54
botocore==1.34.54
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorful==0.5.6
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
exceptiongroup==1.2.0
fastapi==0.108.0
filelock==3.13.1
frozenlist==1.4.1
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gpustat==1.1.1
grpcio==1.62.0
h11==0.14.0
httptools==0.6.1
idna==3.6
importlib-metadata==6.11.0
isort==5.13.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
mccabe==0.7.0
msgpack==1.0.8
multidict==6.0.5
numpy==1.26.4
nvidia-ml-py==12.535.133
oauthlib==3.2.2
ob-metaflow==2.11.4.9
ob-metaflow-extensions==1.1.49
ob-metaflow-stubs==2.11.4.9
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.20.0
opentelemetry-distro==0.41b0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
outerbounds==0.3.58
packaging==23.2
platformdirs==4.2.0
prometheus_client==0.20.0
protobuf==4.25.3
psutil==5.9.8
py-spy==0.3.14
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
pylint==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
ray==2.9.3
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smart-open==7.0.1
sniffio==1.3.1
starlette==0.32.0.post1
threadpoolctl==3.3.0
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
urllib3==2.0.7
uvicorn==0.27.1
uvloop==0.19.0
virtualenv==20.25.1
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
